Cleanup: Move Pagination from uiComponents to common/Pagination
