+++ ui/src/components/common/Pagination/usePagination.ts
@@ -1,7 +1,7 @@
 import { produce } from 'immer';
 import qs from 'qs';
 import { useNavigate } from 'react-router';
-import { useUrlQuery } from '../../hooks/urlQuery/useUrlQuery';
+import { useUrlQuery } from '../../../hooks';
 import { getDisplayedPageNumberList, getRenderedPageNumber } from './utils';
 
 function parseNumberOrDefault(str: string, defaultValue: number): number {
